fix(harness): honour a narrow pane's real width, and keep the log record u64 (#84)
Two fixes, one of them a defect CI caught on the first run.

**The width floor was a made-up number.** `Renderer::new` clamped `cols` up to
20, so every genuinely narrow pane was rendered at a width it never reported —
and `text_deltas_continue_one_line_and_wrap_at_the_column` failed on macOS
saying so (576 other engine tests passed; the module compiled clean). The floor
is now 1, the narrowest real terminal: it exists only so the wrapper cannot be
asked to fit a character into no columns, and every value above it is the
caller's. The zero-width test now pins both halves — a `0` terminates AND is not
widened to a guess — with a real narrow width as the control.

**`LogRecord.ts` is `u64`, not `u128`.** `#[serde(flatten)]` routes the struct
through serde's `Content` buffer, which has no 128-bit variant, so a `u128`
there fails at runtime on a path no type check sees. `obs::now_ms` already
narrows the same way.

+++ crates/loomux-engine/src/harness/mod.rs
@@ -457,7 +457,12 @@ pub const UNKNOWN_LINE_MAX_BYTES: usize = 4096;
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct LogRecord {
     pub seq: u64,
-    pub ts: u128,
+    /// Millis since the epoch, as **`u64`** rather than the `u128`
+    /// `Duration::as_millis` hands back. `#[serde(flatten)]` below routes the
+    /// whole struct through serde's `Content` buffer, which has no 128-bit
+    /// variant — a `u128` here fails at *runtime*, on the one path a type check
+    /// cannot see. `obs::now_ms` already narrows the same way.
+    pub ts: u64,
     #[serde(flatten)]
     pub body: LogBody,
 }
@@ -680,10 +685,10 @@ pub(crate) fn truncate_on_char_boundary(s: &str, max: usize) -> (&str, bool) {
     (&s[..end], true)
 }
 
-fn now_millis() -> u128 {
+fn now_millis() -> u64 {
     std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_millis())
+        .map(|d| d.as_millis() as u64)
         .unwrap_or(0)
 }
